fixed a memory leak in route planner by using shared_ptr refactored parametersmap and corresponding exceptions removed definitively the tcp module

+++ src/18_graph/Journey.cpp
@@ -307,7 +307,7 @@ namespace synthese
 			return _effectiveDuration;
 		}
 
-		int Journey::getDistance() const
+		double Journey::getDistance() const
 		{
 			return _distance;
 		}
